# go-backend/docs/openapi/wave1.yaml
openapi: 3.0.3
info:
  title: POS Bengkel Go API (Wave 1)
  version: 0.1.0
servers:
  - url: http://localhost:8081
paths:
...
  /api/v1/reports/outstanding-payments:
    get:
      summary: Outstanding payments report for completed service orders
      parameters:
        - in: query
          name: page
          schema:
            type: integer
          required: false
      responses:
        '200':
          description: Outstanding payments payload compatible with Laravel Inertia props
...